requirements.txt
google.genterativeai
streamlit
pandas
